Changed Project Strructure for House Keeping
