+++ GP_Framework/__init__.py
@@ -1,0 +1,3 @@
+"""
+Framework package - Contains core GP framework components
+"""
