+++ GP_Indicators/__init__.py
@@ -1,0 +1,3 @@
+"""
+Indicators package - Contains technical analysis indicators and multi-timeframe functions
+"""
