+++ GP_Primitives/__init__.py
@@ -1,0 +1,3 @@
+"""
+Primitives package - Contains basic mathematical and logical operators
+"""
